requirements.txt
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
tensorflow==2.13.0
retina-face==0.0.13
mtcnn==0.1.1
tf-keras==2.13.0
google-generativeai==0.3.2
python-dotenv==1.0.0
streamlit==1.28.1
Pillow==10.0.1